# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "stocker application server.",
        "title": "StockerAPI",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8810",
    "basePath": "/",
    "paths": {
...
        "/products/{janCode}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jancode"
                ],
                "summary": "製品情報取得",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JANコード",
                        "name": "janCode",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.JancodeResponse"
                        }
                    }
                }
            }
        },
...
        "/stores/{storeId}/items": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "item"
                ],
                "summary": "商品一覧取得",
                "parameters": [
                    {
                        "type": "string",
                        "description": "店舗ID",
                        "name": "storeId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.ItemListResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "list": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/controller.ItemResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "item"
                ],
                "summary": "商品新規作成",
                "parameters": [
                    {
                        "description": "商品作成リクエスト",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.ItemRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "店舗ID",
                        "name": "storeId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.ItemResponse"
                        }
                    }
                }
            }
        },
        "/stores/{storeId}/items/{itemId}": {
            "patch": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "item"
                ],
                "summary": "商品更新",
                "parameters": [
                    {
                        "type": "string",
                        "description": "店舗ID",
                        "name": "storeId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "商品ID",
                        "name": "itemId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "商品更新リクエスト",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.ItemRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.ItemResponse"
                        }
                    }
                }
            }
        },
        "/stores/{storeId}/items/{itemId}/stocks": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "item_stock"
                ],
                "summary": "商品詳細取得",
                "parameters": [
                    {
                        "type": "string",
                        "description": "店舗ID",
                        "name": "storeId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "商品ID",
                        "name": "itemId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.ItemStockResponse"
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "item_stock"
                ],
                "summary": "商品詳細登録",
                "parameters": [
                    {
                        "type": "string",
                        "description": "店舗ID",
                        "name": "storeId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "商品ID",
                        "name": "itemId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "商品詳細作成リクエスト",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/controller.ItemStockRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.ItemStockResponse"
                        }
                    }
                }
            },
            "patch": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "item_stock"
                ],
                "summary": "商品詳細更新",
                "parameters": [
                    {
                        "type": "string",
                        "description": "店舗ID",
                        "name": "storeId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "商品ID",
                        "name": "itemId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "商品詳細更新リクエスト",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/controller.ItemStockRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.ItemStockResponse"
                        }
                    }
                }
            }
        },
        "/stores/{storeId}/itemsStocks": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "item_stock"
                ],
                "summary": "商品詳細一覧取得",
                "parameters": [
                    {
                        "type": "string",
                        "description": "店舗ID",
                        "name": "storeId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.ItemStockListResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "list": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/controller.ItemStockResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controller.ItemListResponse": {
            "type": "object",
            "required": [
                "list"
            ],
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controller.ItemResponse"
                    }
                }
            }
        },
        "controller.ItemRequest": {
            "type": "object",
            "required": [
                "janCode",
                "name"
            ],
            "properties": {
                "janCode": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "controller.ItemResponse": {
            "type": "object",
            "required": [
                "createdAt",
                "id",
                "janCode",
                "name",
                "storeId",
                "updatedAt"
            ],
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "janCode": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "storeId": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "controller.ItemStockListResponse": {
            "type": "object",
            "required": [
                "list"
            ],
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controller.ItemStockResponse"
                    }
                }
            }
        },
        "controller.ItemStockRequest": {
            "type": "object",
            "required": [
                "place",
                "stock",
                "stockMin"
            ],
            "properties": {
                "place": {
                    "type": "string"
                },
                "stock": {
                    "type": "integer"
                },
                "stockMin": {
                    "type": "integer"
                }
            }
        },
        "controller.ItemStockResponse": {
            "type": "object",
            "required": [
                "createdAt",
                "itemId",
                "place",
                "stock",
                "stockMin",
                "storeId",
                "updatedAt"
            ],
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "itemId": {
                    "type": "string"
                },
                "place": {
                    "type": "string"
                },
                "stock": {
                    "type": "integer"
                },
                "stockMin": {
                    "type": "integer"
                },
                "storeId": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "controller.JancodeResponse": {
            "type": "object",
            "required": [
                "brandName",
                "makerName",
                "name"
            ],
            "properties": {
                "brandName": {
                    "type": "string"
                },
                "makerName": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    }
}